requirements.txt
pytorch
torchvision
torchaudio
scikit-learn
numpy
seaborn
jupyter
stable-baselines3
gymnasium
